+++ packages/implementations/react/src/composite/Card/src/components/Body/index.ts
@@ -1,0 +1,5 @@
+import { CardBody } from "./Body";
+
+export { CardBody } from "./Body";
+export type { CardBodyProps } from "./body.types";
+export default CardBody;