+++ packages/implementations/react/src/composite/Card/src/index.ts
@@ -1,0 +1,5 @@
+import { Card } from "./Card";
+
+export { Card } from "./Card";
+export type { CardProps } from "./card.types";
+export default Card;